chore(annotations): create utils; move form functions out of reducers into there
