+++ src/annotations/components/annotationForm/AnnotationForm.tsx
@@ -28,7 +28,6 @@ import {
   annotationFormReducer,
   getInitialAnnotationState,
   AnnotationType,
-  annotationFormIsValid,
   Annotation,
   getAnnotationFromDraft,
   AnnotationContextType,
@@ -37,6 +36,10 @@ import {
 
 // Contexts
 import {OverlayContext} from 'src/overlays/components/OverlayController'
+
+import {updateAnnotationDraft} from 'src/annotations/actions/annotationFormActions'
+
+import {annotationFormIsValid} from 'src/annotations/utils/form'
 
 interface Props {
   title: 'Edit' | 'Add'
@@ -82,7 +85,10 @@ export const AnnotationForm: FC<Props> = ({
   const handleSubmit = (e: FormEvent<HTMLFormElement>): void => {
     e.preventDefault()
 
-    const formIsValid = annotationFormIsValid(state, dispatch)
+    // force required inputs into error or valid state (instead of initial)
+    // so user can resolve the form errors and submit again
+    dispatch(updateAnnotationDraft(state))
+    const formIsValid = annotationFormIsValid(state)
 
     if (formIsValid) {
       onSubmit(getAnnotationFromDraft(state))
